requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
pandas==2.2.3g